+++ crates/cli-tools/src/lib.rs
@@ -34,6 +34,10 @@ macro_rules! debug {
 
 #[cfg(feature = "action")]
 pub mod action;
+#[cfg(feature = "cargo")]
+pub mod cargo;
+#[cfg(feature = "changelog")]
+pub mod changelog;
 pub mod cmd;
 pub mod error;
 pub mod fs;